refactor: move boundaries to propper folder
