+++ src/sp-proxy/src/interface-adapters/delivery/CreateRemoteIdpController.ts
@@ -4,7 +4,7 @@ import { IMapper } from '@sp-proxy/interface-adapters/protocols/IMapper'
 import { IRequest } from '@sp-proxy/interface-adapters/protocols/IRequest'
 import { IValidator } from '@sp-proxy/interface-adapters/protocols/IValidator'
 import { CreateRemoteIdpRequestModel } from '@sp-proxy/use-cases/io-models/CreateRemoteIdpRequestModel'
-import { InputBoundary } from '@sp-proxy/use-cases/protocols/InputBoundary'
+import { InputBoundary } from '@sp-proxy/use-cases/io-channels/InputBoundary'
 
 export class CreateRemoteIdpController implements IController {
   constructor(
