+++ src/sp-proxy/src/use-cases/CreateRemoteIdpInteractor.spec.ts
@@ -6,13 +6,13 @@
 import { RemoteIdp } from '@sp-proxy/entities/RemoteIdp'
 import { CreateRemoteIdpInteractor } from '@sp-proxy/use-cases/CreateRemoteIdpInteractor'
 import { makeSingleSignOnServices } from '@sp-proxy/use-cases/factories/makeSingleSignOnServices'
+import { OutputBoundary } from '@sp-proxy/use-cases/io-channels/OutputBoundary'
 import { CreateRemoteIdpRequestModel } from '@sp-proxy/use-cases/io-models/CreateRemoteIdpRequestModel'
 import { CreateRemoteIdpResponseModel } from '@sp-proxy/use-cases/io-models/CreateRemoteIdpResponseModel'
 import { IRequestModel } from '@sp-proxy/use-cases/io-models/IRequestModel'
 import { IResponseModel } from '@sp-proxy/use-cases/io-models/IResponseModel'
 import { ICreateRemoteIdpGateway } from '@sp-proxy/use-cases/ports/ICreateRemoteIdpGateway'
 import { IMapper } from '@sp-proxy/use-cases/protocols/IMapper'
-import { OutputBoundary } from '@sp-proxy/use-cases/protocols/OutputBoundary'
 
 const makePresenter = (): OutputBoundary<CreateRemoteIdpResponseModel> => {
   class PresenterStub implements OutputBoundary<CreateRemoteIdpResponseModel> {
