+++ src/sp-proxy/src/use-cases/io-channels/OutputBoundary.ts
